feat: NumberGeneratorSelector
* Refactored NumberGneratorModal to remove all the selection logic, and instead moved that to its own exposed component.
* This component renders a QueryTypedown and can now be directly implemented as the user sees fit
* Moved endpoint constants from "utilities" to "constants"
