+++ src/public/components/NumberGeneratorSelector/index.js
@@ -1,0 +1,5 @@
+export {
+  default as NumberGeneratorSelector,
+  SEQUENCE_TYPEDOWN_ID,
+  SEQUENCE_TYPEDOWN_ID_UNIQUE
+} from './NumberGeneratorSelector';
