+++ src/public/components/index.js
@@ -4,3 +4,8 @@ export {
   default as NumberGeneratorModal,
   LegacyNumberGeneratorModal
 } from './NumberGeneratorModal';
+export {
+  NumberGeneratorSelector,
+  SEQUENCE_TYPEDOWN_ID,
+  SEQUENCE_TYPEDOWN_ID_UNIQUE
+} from './NumberGeneratorSelector';
