+++ src/public/constants/index.js
@@ -1,1 +1,2 @@
 export * from './maximumCheckValues';
+export * from './endpoints';
