--- src/public/utilities/endpoints.js
@@ -1,5 +1,0 @@
-export const NUMBER_GENERATORS_ENDPOINT = 'servint/numberGenerators';
-export const NUMBER_GENERATOR_ENDPOINT = (id) => `${NUMBER_GENERATORS_ENDPOINT}/${id}`;
-
-export const NUMBER_GENERATOR_SEQUENCES_ENDPOINT = 'servint/numberGeneratorSequences';
-export const NUMBER_GENERATOR_SEQUENCE_ENDPOINT = (id) => `${NUMBER_GENERATOR_SEQUENCES_ENDPOINT}/${id}`;
