+++ src/public/utilities/index.js
@@ -1,4 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as ModalButton } from './ModalButton';
-
-export * from './endpoints';
